Added MonsterParty, Monster Bundles in monster.rs
Current Implementation focuses on Bundles because stats should equal components, and other components to keep track of things.

Further reworkings into Enemies, Moves, and displaying the Monsters will need to be taken into consideration, this just includes essential monster data into the game so that we have something that is a Monster in the game's memory.